Fix SwiftLint violations and improve code organization
Resolves type contents order, multiline formatting, and file length violations across the codebase. Splits oversized test file for better maintainability and extracts ImageDictionary extension for improved modularity.
